requirements.txt
msal>=1.30.0
requests>=2.32.0
PyYAML>=6.0.2
